refactor(live-v2): factor shared session logic into helpers
Extract duplicated sync/async live v2 session code into _helpers.py:
chunked audio resend, ws message parsing, lifecycle events, and the
typed event listener mixin.
